Update insert.md
I think there is a bug in the docs here. The second argument to the lambda for `onConflictUpdateSet` accesses the excluded row (the one you were attempting to insert) not the old value. From [https://github.com/haskell-beam/beam/blob/master/beam-postgres/Database/Beam/Postgres/Full.hs#L457](here), it looks like the correct way to access the old value is `current_ (customerEmail Fields)`, which also behaves as expected when I try it out in my code. 

I haven't actually tested this update (I tested the equivalent in my code), but a PR pointing to this line seemed like the most efficient way to convey the bug report.